fix: Merge error and some refactoring
