+++ apis/core/lib/domain/dimension/DimensionMetadataCollection.ts
@@ -1,11 +1,22 @@
 import { GraphPointer } from 'clownface'
 import { Constructor } from '@tpluscode/rdfine'
 import { cc } from '@cube-creator/core/namespace'
-import { DimensionMetadataCollection } from '@cube-creator/model'
+import { ColumnMapping, CsvMapping, DimensionMetadataCollection } from '@cube-creator/model'
 import { schema } from '@tpluscode/rdf-ns-builders'
+import * as DimensionMetadata from '@cube-creator/model/DimensionMetadata'
+import $rdf from 'rdf-ext'
+import { shrink } from '@zazuko/rdf-vocabularies'
+import { ResourceStore } from '../../ResourceStore'
+
+interface CreateColumnMetadata {
+  store: ResourceStore
+  csvMapping: CsvMapping
+  columnMapping: ColumnMapping
+}
 
 interface ApiDimensionMetadataCollection {
   updateDimension(dimension: GraphPointer): void
+  addDimensionMetadata(params: CreateColumnMetadata): DimensionMetadata.DimensionMetadata
 }
 
 declare module '@cube-creator/model' {
@@ -26,6 +37,30 @@ export default function Mixin<Base extends Constructor<Omit<DimensionMetadataCol
       for (const quad of dimension.dataset.match(dimension.term)) {
         found.pointer.addOut(quad.predicate, quad.object)
       }
+    }
+
+    addDimensionMetadata(params: CreateColumnMetadata): DimensionMetadata.DimensionMetadata {
+      const identifier = params.csvMapping.createIdentifier(params.columnMapping.targetProperty)
+
+      if (this.hasPart.some((dimMeta) => dimMeta.about === identifier)) {
+        throw new Error(`Dimension Metadata with identifier ${identifier.value} already exists`)
+      }
+
+      const dimensionMetadata = DimensionMetadata.create(this.pointer.node(this.getId(params.columnMapping)), {
+        about: identifier,
+      })
+
+      this.pointer.addOut(schema.hasPart, dimensionMetadata.id)
+
+      return dimensionMetadata
+    }
+
+    private getId(mapping: ColumnMapping) {
+      const property = (mapping.targetProperty.termType === 'Literal')
+        ? mapping.targetProperty.value
+        : shrink(mapping.targetProperty.value) || encodeURI(mapping.targetProperty.value)
+
+      return $rdf.namedNode(`${this.id.value}/${property}`)
     }
   }
 }
