+++ apis/core/lib/domain/index.ts
@@ -8,7 +8,6 @@ import Table from './table/Table'
 import Dataset from './dataset/Dataset'
 import Resource from './Resource'
 import DimensionMetadataCollection from './dimension/DimensionMetadataCollection'
-import DimensionMetadata from './dimension-metadata/DimensionMetadata'
 
 RdfResourceImpl.factory.addMixin(...Object.values(CoreModel))
 RdfResourceImpl.factory.addMixin(CsvMapping)
@@ -18,6 +17,5 @@ RdfResourceImpl.factory.addMixin(Dataset)
 RdfResourceImpl.factory.addMixin(Table)
 RdfResourceImpl.factory.addMixin(Resource)
 RdfResourceImpl.factory.addMixin(DimensionMetadataCollection)
-RdfResourceImpl.factory.addMixin(DimensionMetadata)
 
 RdfResourceImpl.factory.addMixin(ThingMixin)
